[BE] feat: jql 수정

See merge request s13-ai-image-sub1/S13P21E101!96

+++ FE/ConvenienceSightApp/app/src/main/java/com/example/lookey/data/model/allergy/AllergyAddResponse.kt
@@ -10,6 +10,6 @@ data class AllergyPostRequest(
     val request: Request
 ) {
     data class Request(
-        val allergy_id: Int
+        val allergy_id: Long
     )
 }
